Added SMC traps for ARM64 platforms
This patch adds architectur preproceesor to differentiate x86 and ARM64.
This patch further adds an smc_cb which is similar to the int3_cb.
We also introduce sw_trap as a name for both software traps:
INT3 and SMC.

Because this patches uses inject_trap_breakpoint in the code path,
we renamed the function to inject_trap_sw.

The few changes needed to support ARM32 are going to be tested and submitted
next week.

This patch need the libvmi commit: 69b3198 to build.